# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PayD Backend API",
    "version": "1.0.0",
    "description": "API Documentation for PayD Backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/contracts/{registryId}": {
      "get": {
        "summary": "Single contract detail",
        "tags": [
          "Contract Upgrades"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "registryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/contracts/{registryId}/validate-hash": {
      "post": {
        "summary": "Validates WASM hash format and on-chain existence",
        "tags": [
          "Contract Upgrades"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "registryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newWasmHash": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/contracts/{registryId}/simulate-upgrade": {
      "post": {
        "summary": "Pre-flights the upgrade via Soroban RPC",
        "tags": [
          "Contract Upgrades"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "registryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newWasmHash": {
                    "type": "string"
                  },
                  "initiatedBy": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/contracts/{registryId}/upgrade-logs": {
      "get": {
        "summary": "Paginated upgrade history for a specific contract",
        "tags": [
          "Contract Upgrades"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "registryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/contracts/upgrade-logs/{logId}/execute": {
      "post": {
        "summary": "Executes a simulated upgrade on-chain",
        "tags": [
          "Contract Upgrades"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "logId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "adminSecret": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/contracts/upgrade-logs/{logId}/status": {
      "get": {
        "summary": "Polls migration step progress for an executing upgrade",
        "tags": [
          "Contract Upgrades"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "logId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/contracts/upgrade-logs/{logId}/cancel": {
      "post": {
        "summary": "Cancels a pending or simulated upgrade before execution",
        "tags": [
          "Contract Upgrades"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "logId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/employees": {
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "summary": "Get all employees",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/employees/{id}": {
      "get": {
        "summary": "Get a single employee",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "summary": "Update an employee",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "summary": "Delete an employee",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/employees/bulk-import": {
      "post": {
        "summary": "Bulk import employees from CSV",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/payroll/employees/{employeeId}": {
      "get": {
        "summary": "Get payroll for a specific employee",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "employeeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "orgPublicKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/payroll/audit/employee/{employeeId}": {
      "get": {
        "summary": "Get audit logs for a specific employee",
        "tags": [
          "Payroll Audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "employeeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/search/organizations/{organizationId}/employees": {
      "get": {
        "summary": "Search and filter employees",
        "tags": [
          "Data Search"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/trustline/employees/{employeeId}": {
      "get": {
        "summary": "Get stored trustline status for an employee",
        "tags": [
          "Trustlines"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "employeeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}